chore: repurpose demo debug screen as teams
